requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
tensorflow>=2.10.0
opencv-python>=4.5.0
pillow>=8.0.0
numpy>=1.21.0
pandas>=1.3.0
requests>=2.25.0
beautifulsoup4>=4.9.0
scikit-learn>=1.0.0
matplotlib>=3.3.0
seaborn>=0.11.0
plotly>=5.0.0
python-dotenv>=0.19.0
